+++ Express/package.json
@@ -13,7 +13,6 @@
     "aws-sdk": "^2.1613.0",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
-    "mssql": "^10.0.2",
-    "nodemon": "^3.1.0"
+    "mssql": "^10.0.2"
   }
 }
